Refactored website so data is not in live
Now there is a nongenerated website which contains all the html and
assets which are not generated. The live website, hosted on aws, is
removed from version control. A new script has been made to maintain it
which copies over the nongenerated files which are hosted as is and
calls the python scripts which makes the blog and index, the only
currently generated files.