Mobile CSS
Changed the css to look more reasonable on mobile. On mobile all menus
are now on the top and they are hidden on scrolling down. The css is
also generally cleaned up but still not well. TODO: prune down on the
menu namespace and think of what you actually need.