Major Refactor
The general philosophy of the site has become more streamlines and
consistent. Now all webpages have corresponding json files in the data
directory which point to appropriate template and content files. A
notion of component is created so now the sidebar and header will not
need to be modified in many places and the css version can be updated
more conveniently.